restructure and fixes from xsd-compliant branch
