+++ pkg/utils/xml.go
@@ -1,9 +1,6 @@
-package wms130
+package utils
 
-import (
-	"encoding/xml"
-	"fmt"
-)
+import "encoding/xml"
 
 // XMLAttribute wrapper around the array of xml.Attr
 type XMLAttribute []xml.Attr
@@ -40,36 +37,6 @@ func (xmlattr *XMLAttribute) UnmarshalXML(d *xml.Decoder, start xml.StartElement
 		switch el := token.(type) {
 		case xml.EndElement:
 			if el == start.End() {
-				return nil
-			}
-		}
-	}
-}
-
-// MarshalXML Position
-func (c *CRS) MarshalXML(e *xml.Encoder, start xml.StartElement) error {
-	var s = ``
-	if c.Namespace != `` {
-		s = fmt.Sprintf("%s:%d", c.Namespace, c.Code)
-	}
-
-	return e.EncodeElement(s, start)
-}
-
-// UnmarshalXML Position
-func (c *CRS) UnmarshalXML(d *xml.Decoder, start xml.StartElement) error {
-	var crs CRS
-	for {
-		token, err := d.Token()
-		if err != nil {
-			return err
-		}
-		switch el := token.(type) {
-		case xml.CharData:
-			crs.parseString(string([]byte(el)))
-		case xml.EndElement:
-			if el == start.End() {
-				*c = crs
 				return nil
 			}
 		}
